+++ gulp/build.js
@@ -1,5 +1,6 @@
 'use strict';
 
 var gulp = require('gulp');
+var rump = require('../index');
 
-gulp.task('rump:build', ['rump:clean']);
+gulp.task(rump.taskName('build'), [rump.taskName('clean')]);