+++ gulp/clean.js
@@ -3,7 +3,8 @@
 var del = require('del');
 var gulp = require('gulp');
 var configs = require('../configs');
+var rump = require('../index');
 
-gulp.task('rump:clean', function() {
+gulp.task(rump.taskName('clean'), function() {
   del.sync(configs.main.paths.destination.root);
 });