+++ gulp/watch.js
@@ -2,10 +2,14 @@
 
 var gulp = require('gulp');
 var configs = require('../configs');
+var rump = require('../index');
 configs.watch = false;
 
-gulp.task('rump:watch', ['rump:clean', 'rump:watch:setup']);
+gulp.task(rump.taskName('watch'), [
+  rump.taskName('clean'),
+  rump.taskName('watch:setup')
+]);
 
-gulp.task('rump:watch:setup', function() {
+gulp.task(rump.taskName('watch:setup'), function() {
   configs.watch = true;
 });